requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
joblib
paho-mqtt
cryptography
pyjwt
pytest
python-dotenv
elasticsearch # For connecting to Elasticsearch later
azure-identity
azure-mgmt-network